Move util.py, interface.py -> common package
