+++ common/interface.py
